cleanup in enriching module
