+++ envinorma/parametrization/am_with_versions.py
@@ -2,8 +2,10 @@ from dataclasses import dataclass, replace
 from datetime import date
 from typing import Dict, Optional, Tuple
 
-from envinorma.am_enriching import add_table_inspection_sheet_data, detect_and_add_topics, remove_null_applicabilities
+from envinorma.enriching.remove_null_applicabilities import remove_null_applicabilities
+from envinorma.enriching.table_row_inline_content import add_table_inspection_sheet_data
 from envinorma.enriching.title_reference import add_references
+from envinorma.enriching.topic_detection import detect_and_add_topics
 from envinorma.models.am_metadata import AMMetadata
 from envinorma.models.arrete_ministeriel import ArreteMinisteriel, add_metadata
 from envinorma.models.classement import ClassementWithAlineas, Regime
